pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "localbrain"
dynamic = ["version"]
description = "Local Brain - Personal knowledge management system for collecting, processing, and querying knowledge"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Agentic Local Brain Team"}
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "pypinyin>=0.49",
    # Web collectors
    "httpx>=0.24.0",
    "readability-lxml>=0.8.1",
    "markdownify>=0.11.0",
    "beautifulsoup4>=4.12.0",
    # PDF processing
    "PyPDF2>=3.0.0",
    # AI/ML
    "dashscope>=1.14.0",
    "openai>=1.0.0",
    "litellm>=1.30.0",
    # Knowledge Mining
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "scikit-learn>=1.3.0",
    # Vector storage
    "chromadb>=0.4.0",
    # Web API
    "fastapi>=0.95.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.5",
]

[project.scripts]
localbrain = "kb.cli:cli"
kb = "kb.cli:cli"

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]
build = [
    "pyinstaller>=5.0",
    "pyinstaller-hooks-contrib>=2023.0",
]

[tool.setuptools.packages.find]
include = ["kb*"]

[tool.setuptools.package-data]
kb = [
    "config-template.yaml",
    "web/static/**/*",
    "web/static/*",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (run separately)",
    "slow: marks tests as slow running",
]
addopts = "--ignore=tests/test_model_services_integration.py --ignore=tests/test_vectorization_and_tagging.py"
